Remove array properties from types.
