+++ include/dynd/kernels/field_access_kernel.hpp
@@ -63,6 +63,60 @@ namespace nd {
     }
   };
 
+  // Temporary solution (previously in struct_type.cpp).
+  struct get_array_field_kernel : nd::base_kernel<get_array_field_kernel> {
+    array self;
+    intptr_t i;
+
+    get_array_field_kernel(const array &self, intptr_t i) : self(self), i(i) {}
+
+    void call(array *dst, const array *DYND_UNUSED(src))
+    {
+      array res = helper(self, i);
+      *dst = res;
+    }
+
+    static void resolve_dst_type(char *static_data, char *DYND_UNUSED(data), ndt::type &dst_tp,
+                                 intptr_t DYND_UNUSED(nsrc), const ndt::type *DYND_UNUSED(src_tp),
+                                 intptr_t DYND_UNUSED(nkwd), const array *kwds,
+                                 const std::map<std::string, ndt::type> &DYND_UNUSED(tp_vars))
+    {
+      dst_tp = helper(kwds[0], *reinterpret_cast<intptr_t *>(static_data)).get_type();
+    }
+
+    static void instantiate(char *static_data, char *DYND_UNUSED(data), nd::kernel_builder *ckb,
+                            const ndt::type &DYND_UNUSED(dst_tp), const char *DYND_UNUSED(dst_arrmeta),
+                            intptr_t DYND_UNUSED(nsrc), const ndt::type *DYND_UNUSED(src_tp),
+                            const char *const *DYND_UNUSED(src_arrmeta), kernel_request_t kernreq,
+                            intptr_t DYND_UNUSED(nkwd), const array *kwds,
+                            const std::map<std::string, ndt::type> &DYND_UNUSED(tp_vars))
+    {
+      ckb->emplace_back<get_array_field_kernel>(kernreq, kwds[0], *reinterpret_cast<intptr_t *>(static_data));
+    }
+
+    static array helper(const array &n, intptr_t i)
+    {
+      // Get the nd::array 'self' parameter
+      intptr_t undim = n.get_ndim();
+      ndt::type udt = n.get_dtype();
+      if (udt.get_kind() == expr_kind) {
+        std::string field_name = udt.value_type().extended<ndt::struct_type>()->get_field_name(i);
+        return n.replace_dtype(ndt::make_type<ndt::adapt_type>(
+            udt.value_type().extended<ndt::struct_type>()->get_field_type(i), udt, nd::callable(), nd::callable()));
+      }
+      else {
+        if (undim == 0) {
+          return n(i);
+        }
+        else {
+          shortvector<irange> idx(undim + 1);
+          idx[undim] = irange(i);
+          return n.at_array(undim + 1, idx.get());
+        }
+      }
+    }
+  };
+
 } // namespace dynd::nd
 
 namespace ndt {
