+++ include/dynd/types/adapt_type.hpp
@@ -38,7 +38,6 @@ namespace ndt {
 
     bool operator==(const base_type &rhs) const;
 
-    std::map<std::string, nd::callable> get_dynamic_array_properties() const;
     std::map<std::string, nd::callable> get_dynamic_array_functions() const;
   };
