+++ include/dynd/types/categorical_type.hpp
@@ -89,7 +89,6 @@ namespace ndt {
     void arrmeta_debug_print(const char *arrmeta, std::ostream &o, const std::string &indent) const;
 
     std::map<std::string, nd::callable> get_dynamic_type_properties() const;
-    std::map<std::string, nd::callable> get_dynamic_array_properties() const;
 
     friend struct assign_to_same_category_type;
     friend struct assign_from_same_category_type;
