+++ src/dynd/types/fixed_dim_kind_type.cpp
@@ -212,11 +212,6 @@ std::map<std::string, nd::callable> ndt::fixed_dim_kind_type::get_dynamic_type_p
   return properties;
 }
 
-std::map<std::string, nd::callable> ndt::fixed_dim_kind_type::get_dynamic_array_properties() const
-{
-  return m_element_tp.get_array_properties();
-}
-
 std::map<std::string, nd::callable> ndt::fixed_dim_kind_type::get_dynamic_array_functions() const
 {
   return m_element_tp.get_array_functions();
