+++ src/dynd/types/fixed_dim_type.cpp
@@ -24,9 +24,6 @@ ndt::fixed_dim_type::fixed_dim_type(intptr_t dim_size, const type &element_tp)
   this->flags |= (element_tp.get_flags() & (type_flags_operand_inherited | type_flags_value_inherited));
 
   if ((get_flags() & type_flag_symbolic) == 0) {
-    std::map<std::string, nd::callable> element_properties = element_tp.get_array_properties();
-    m_array_properties.insert(element_properties.begin(), element_properties.end());
-
     std::map<std::string, nd::callable> element_functions = element_tp.get_array_functions();
     m_array_functions.insert(element_functions.begin(), element_functions.end());
   }
@@ -631,11 +628,6 @@ std::map<std::string, nd::callable> ndt::fixed_dim_type::get_dynamic_type_proper
       [](type self) { return type(self.extended<fixed_dim_type>()->get_element_type()); }, "self");
 
   return properties;
-}
-
-std::map<std::string, nd::callable> ndt::fixed_dim_type::get_dynamic_array_properties() const
-{
-  return m_array_properties;
 }
 
 std::map<std::string, nd::callable> ndt::fixed_dim_type::get_dynamic_array_functions() const
