+++ src/dynd/types/struct_type.cpp
@@ -40,8 +40,6 @@ ndt::struct_type::struct_type(const std::vector<std::string> &field_names, const
   }
 
   this->kind = variadic ? kind_kind : struct_kind;
-
-  create_array_properties();
 }
 
 ndt::struct_type::~struct_type() {}
@@ -320,104 +318,6 @@ std::map<std::string, nd::callable> ndt::struct_type::get_dynamic_type_propertie
                                ndt::struct_type::make({"self"}, {ndt::make_type<ndt::type_type>()})));
 
   return properties;
-}
-
-namespace dynd {
-namespace nd {
-
-  struct get_array_field_kernel : nd::base_kernel<get_array_field_kernel> {
-    array self;
-    intptr_t i;
-
-    get_array_field_kernel(const array &self, intptr_t i) : self(self), i(i) {}
-
-    void call(array *dst, const array *DYND_UNUSED(src))
-    {
-      array res = helper(self, i);
-      *dst = res;
-    }
-
-    static void resolve_dst_type(char *static_data, char *DYND_UNUSED(data), ndt::type &dst_tp,
-                                 intptr_t DYND_UNUSED(nsrc), const ndt::type *DYND_UNUSED(src_tp),
-                                 intptr_t DYND_UNUSED(nkwd), const array *kwds,
-                                 const std::map<std::string, ndt::type> &DYND_UNUSED(tp_vars))
-    {
-      dst_tp = helper(kwds[0], *reinterpret_cast<intptr_t *>(static_data)).get_type();
-    }
-
-    static void instantiate(char *static_data, char *DYND_UNUSED(data), nd::kernel_builder *ckb,
-                            const ndt::type &DYND_UNUSED(dst_tp), const char *DYND_UNUSED(dst_arrmeta),
-                            intptr_t DYND_UNUSED(nsrc), const ndt::type *DYND_UNUSED(src_tp),
-                            const char *const *DYND_UNUSED(src_arrmeta), kernel_request_t kernreq,
-                            intptr_t DYND_UNUSED(nkwd), const array *kwds,
-                            const std::map<std::string, ndt::type> &DYND_UNUSED(tp_vars))
-    {
-      ckb->emplace_back<get_array_field_kernel>(kernreq, kwds[0], *reinterpret_cast<intptr_t *>(static_data));
-    }
-
-    static array helper(const array &n, intptr_t i)
-    {
-      // Get the nd::array 'self' parameter
-      intptr_t undim = n.get_ndim();
-      ndt::type udt = n.get_dtype();
-      if (udt.get_kind() == expr_kind) {
-        std::string field_name = udt.value_type().extended<ndt::struct_type>()->get_field_name(i);
-        return n.replace_dtype(ndt::make_type<ndt::adapt_type>(
-            udt.value_type().extended<ndt::struct_type>()->get_field_type(i), udt, nd::callable(), nd::callable()));
-      }
-      else {
-        if (undim == 0) {
-          return n(i);
-        }
-        else {
-          shortvector<irange> idx(undim + 1);
-          idx[undim] = irange(i);
-          return n.at_array(undim + 1, idx.get());
-        }
-      }
-    }
-  };
-
-} // namespace dynd::nd
-
-} // namespace dynd
-
-static std::vector<ndt::type> make_self_types()
-{
-  std::vector<ndt::type> result(1);
-  result[0] = ndt::any_kind_type::make();
-  return result;
-}
-
-ndt::struct_type::struct_type(int, int) : tuple_type(struct_id, make_self_types(), type_flag_none, false, false)
-{
-  // Equivalent to ndt::struct_type::make(ndt::make_ndarrayarg(), "self");
-  // but hardcoded to break the dependency of struct_type::array_parameters_type
-  m_arrmeta_offsets = {0};
-  // The data offsets also consist of one zero
-  //    m_data_offsets = m_arrmeta_offsets;
-  // Inherit any operand flags from the fields
-  this->flags |= (ndt::any_kind_type::make().get_flags() & type_flags_operand_inherited);
-  this->data_alignment = sizeof(void *);
-  this->arrmeta_size = 0;
-  this->data_size = sizeof(void *);
-  // Leave m_array_properties so there is no reference loop
-}
-
-void ndt::struct_type::create_array_properties()
-{
-  type array_parameters_type(new struct_type(0, 0), false);
-
-  for (intptr_t i = 0, i_end = m_field_count; i != i_end; ++i) {
-    // TODO: Transform the name into a valid Python symbol?
-    m_array_properties[get_field_name(i)] = nd::callable::make<nd::get_array_field_kernel>(
-        callable_type::make(type("Any"), tuple_type::make(), array_parameters_type), i);
-  }
-}
-
-std::map<std::string, nd::callable> ndt::struct_type::get_dynamic_array_properties() const
-{
-  return m_array_properties;
 }
 
 bool ndt::struct_type::match(const char *arrmeta, const type &candidate_tp, const char *candidate_arrmeta,
